p4d restore check for linux

+++ cli_p4d_restore.sh
@@ -1,3 +1,7 @@
-rm -rf data/db.*
-rm -rf data/journal
-./cli_p4d.sh -r data -jr checkpoint.1
+if [ -f data/checkpoint.$1 ]; then
+    rm -rf data/db.*
+    rm -rf data/journal
+    ./cli_p4d.sh -r data -jr checkpoint.$1
+else
+    echo Not found data/checkpoint.$1
+fi
